Ошибки исправлены

+++ lab/lab1/b/ArrayMass.java
@@ -11,42 +11,42 @@ import java.util.Scanner;
 public class ArrayMass {
     static Scanner input = new Scanner(System.in);
     public static void main (String [] args){
-        int [] a = inputArray();
-        int [] b = inputArray();
+        byte [] a = inputArray();
+        byte [] b = inputArray();
         printArray(mutippleArrays(a,b));
     }
-    static int [] mutippleArrays(int [] firstArray,int [] secondArray){
+    static byte [] mutippleArrays(byte[] firstArray,byte[] secondArray){
         int resultNumber = convertToNumberArray(firstArray) * convertToNumberArray(secondArray);
         return convertNumberToArray(resultNumber);
     }
-    static int convertToNumberArray(int[] array){
+    static int convertToNumberArray(byte[] array){
         int number = 0;
         for (int i= array.length-1; i >= 0; i--){
             number += array[i] * (int) Math.pow(10,array.length - i-1);
         }
         return number;
     }
-    static int [] convertNumberToArray(int number){
+    static byte [] convertNumberToArray(int number){
         int arrayLength = ("" + number).length();
-        int [] result = new int [arrayLength];
+        byte [] result = new byte [arrayLength];
         for (int i= arrayLength-1; i >= 0; i--){
-            result[i] = number%10;
+            result[i] = (byte) (number%10);
             number/=10;
         }
         return result;
     }
 
-    static int[] inputArray(){
+    static byte[] inputArray(){
         System.out.print("Введите длину массива:");
         int arrayLength = input.nextInt();
-        int [] array = new int [arrayLength];
+        byte [] array = new byte [arrayLength];
         System.out.print("Введите элементы массива:");
         for(int i=0; i< arrayLength; i++){
-            array[i] = input.nextInt();
+            array[i] = input.nextByte();
         }
         return array;
     }
-    static void printArray(int [] array){
+    static void printArray(byte [] array){
         System.out.print("[");
         for (int i=0; i< array.length; i++){
             System.out.print(array[i]+",");
